+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - community: proffered plugins from astrocommunity
 - colorscheme: load and set catppuccin as theme
 - snippets: include vscode snippets from practicalli astro
+- practicalli: neovim & plugin preferences, custom key maps
 
 ## Changed
 - lazy: enable zipPlugin to view src inside jar files